Manipulator works now :)

+++ libs/CSMUtil/CSMUtil.hpp
@@ -1,10 +1,13 @@
 #ifndef CSMUTIL_HPP
 #define CSMUTIL_HPP
 
+#include <bitset>
+#include <iostream>
+
 namespace csmutil{
 
 static int imap(float value, float from_min, float from_max, int to_min, int to_max){
-    return (int)((value - from_min) * (to_max - to_min) / (from_max - from_min) + to_min);
+    return ((value - from_min) * ((to_max - to_min) / (from_max - from_min)) + to_min);
 }
 
 template<typename R, typename T, typename F>
